pyproject.toml
[tool.poetry]
name = "shaman"
version = "0.1.0"
description = "University course scheduler and planner backend"
authors = ["Cesar Ramirez <cesar.i.ramirez02@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
openpyxl = "^3.1.2"
gurobipy = "*"
python-dotenv = "^1.0.1"
deepseek = "^1.0.0"
transformers = "^4.50.3"
torch = "^2.6.0"
accelerate = "^0.26.0"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
